+++ src/backend/joanie/tests/core/test_models_order_get_or_generate_certificate_for_certificate_product.py
@@ -25,16 +25,19 @@ class CertificateProductGetOrGenerateCertificateOrderModelsTestCase(TestCase):
             enrollment_end=timezone.now() + timedelta(hours=1),
             enrollment_start=timezone.now() - timedelta(hours=1),
             is_gradable=True,
+            is_listed=True,
             start=timezone.now() - timedelta(hours=1),
         )
         product = factories.ProductFactory(
             price="0.00",
             type="certificate",
             certificate_definition=factories.CertificateDefinitionFactory(),
-            target_courses=[course_run.course],
+            courses=[course_run.course],
         )
-        factories.CourseFactory(products=[product])
-        order = factories.OrderFactory(product=product)
+        order = factories.OrderFactory(product=product, course=course_run.course)
+        factories.EnrollmentFactory(
+            user=order.owner, course_run=course_run, is_active=True
+        )
 
         new_certificate, created = order.get_or_generate_certificate()
 
@@ -74,48 +77,19 @@ class CertificateProductGetOrGenerateCertificateOrderModelsTestCase(TestCase):
             enrollment_end=timezone.now() + timedelta(hours=1),
             enrollment_start=timezone.now() - timedelta(hours=1),
             is_gradable=True,
+            is_listed=True,
             start=timezone.now() - timedelta(hours=1),
         )
         product = factories.ProductFactory(
             price="0.00",
             type="certificate",
             certificate_definition=None,  # No certificate definition defined
-            target_courses=[course_run.course],
+            courses=[course_run.course],
         )
-        factories.CourseFactory(products=[product])
-        order = factories.OrderFactory(product=product)
-
-        new_certificate, created = order.get_or_generate_certificate()
-
-        self.assertFalse(created)
-        self.assertFalse(Certificate.objects.exists())
-        self.assertIsNone(new_certificate)
-
-    def test_models_order_get_or_generate_certificate_for_certificate_product_no_graded_course(
-        self,
-    ):
-        """
-        No certificate should be generated if the product does not have at least one
-        graded course.
-        """
-        course_run = factories.CourseRunFactory(
-            enrollment_end=timezone.now() + timedelta(hours=1),
-            enrollment_start=timezone.now() - timedelta(hours=1),
-            is_gradable=True,
-            start=timezone.now() - timedelta(hours=1),
+        order = factories.OrderFactory(product=product, course=course_run.course)
+        factories.EnrollmentFactory(
+            user=order.owner, course_run=course_run, is_active=True
         )
-        product = factories.ProductFactory(
-            price="0.00",
-            type="certificate",
-            certificate_definition=factories.CertificateDefinitionFactory(),
-        )
-        factories.ProductTargetCourseRelationFactory(
-            product=product,
-            course=course_run.course,
-            is_graded=False,  # course is not graded
-        )
-        factories.CourseFactory(products=[product])
-        order = factories.OrderFactory(product=product)
 
         new_certificate, created = order.get_or_generate_certificate()
 
@@ -126,21 +100,24 @@ class CertificateProductGetOrGenerateCertificateOrderModelsTestCase(TestCase):
     def test_models_order_get_or_generate_certificate_for_certificate_product_run_is_not_gradable(
         self,
     ):
-        """No certificate should be generated if there is no gradable course runs."""
+        """No certificate should be generated if the courserun is not gradable."""
         course_run = factories.CourseRunFactory(
             enrollment_end=timezone.now() + timedelta(hours=1),
             enrollment_start=timezone.now() - timedelta(hours=1),
             is_gradable=False,  # course run is not gradable
+            is_listed=True,
             start=timezone.now() - timedelta(hours=1),
         )
         product = factories.ProductFactory(
             price="0.00",
             type="certificate",
             certificate_definition=factories.CertificateDefinitionFactory(),
-            target_courses=[course_run.course],
+            courses=[course_run.course],
         )
-        factories.CourseFactory(products=[product])
-        order = factories.OrderFactory(product=product)
+        order = factories.OrderFactory(product=product, course=course_run.course)
+        factories.EnrollmentFactory(
+            user=order.owner, course_run=course_run, is_active=True
+        )
 
         new_certificate, created = order.get_or_generate_certificate()
 
@@ -151,11 +128,12 @@ class CertificateProductGetOrGenerateCertificateOrderModelsTestCase(TestCase):
     def test_models_order_get_or_generate_certificate_for_certificate_product_run_in_the_future(
         self,
     ):
-        """No certificate should be generated if the course runs are in the future."""
+        """No certificate should be generated if the course run is in the future."""
         course_run = factories.CourseRunFactory(
             enrollment_end=timezone.now() + timedelta(hours=1),
             enrollment_start=timezone.now() - timedelta(hours=1),
             is_gradable=True,
+            is_listed=True,
             start=timezone.now()
             + timedelta(hours=1),  # course run starts in the future
         )
@@ -163,10 +141,12 @@ class CertificateProductGetOrGenerateCertificateOrderModelsTestCase(TestCase):
             price="0.00",
             type="certificate",
             certificate_definition=factories.CertificateDefinitionFactory(),
-            target_courses=[course_run.course],
+            courses=[course_run.course],
         )
-        factories.CourseFactory(products=[product])
-        order = factories.OrderFactory(product=product)
+        order = factories.OrderFactory(product=product, course=course_run.course)
+        factories.EnrollmentFactory(
+            user=order.owner, course_run=course_run, is_active=True
+        )
 
         new_certificate, created = order.get_or_generate_certificate()
 
@@ -182,19 +162,19 @@ class CertificateProductGetOrGenerateCertificateOrderModelsTestCase(TestCase):
             enrollment_end=timezone.now() + timedelta(hours=1),
             enrollment_start=timezone.now() - timedelta(hours=1),
             is_gradable=True,
+            is_listed=True,
             start=timezone.now() - timedelta(hours=1),
         )
         product = factories.ProductFactory(
             price="0.00",
             type="certificate",
             certificate_definition=factories.CertificateDefinitionFactory(),
-            target_courses=[course_run.course],
+            courses=[course_run.course],
         )
-        factories.CourseFactory(products=[product])
-        order = factories.OrderFactory(product=product)
-        enrollment = Enrollment.objects.get()
-        enrollment.is_active = False
-        enrollment.save()
+        order = factories.OrderFactory(product=product, course=course_run.course)
+        factories.EnrollmentFactory(
+            user=order.owner, course_run=course_run, is_active=False
+        )
 
         new_certificate, created = order.get_or_generate_certificate()
 
@@ -210,16 +190,19 @@ class CertificateProductGetOrGenerateCertificateOrderModelsTestCase(TestCase):
             enrollment_end=timezone.now() + timedelta(hours=1),
             enrollment_start=timezone.now() - timedelta(hours=1),
             is_gradable=True,
+            is_listed=True,
             start=timezone.now() - timedelta(hours=1),
         )
         product = factories.ProductFactory(
             price="0.00",
             type="certificate",
             certificate_definition=factories.CertificateDefinitionFactory(),
-            target_courses=[course_run.course],
+            courses=[course_run.course],
         )
-        factories.CourseFactory(products=[product])
-        order = factories.OrderFactory(product=product)
+        order = factories.OrderFactory(product=product, course=course_run.course)
+        factories.EnrollmentFactory(
+            user=order.owner, course_run=course_run, is_active=True
+        )
 
         with mock.patch.object(Enrollment, "get_grade", return_value={"passed": False}):
             new_certificate, created = order.get_or_generate_certificate()